+++ frontend/__tests__/components/file-explorer/explorer-tree.test.tsx
@@ -1,7 +1,7 @@
 import { screen } from "@testing-library/react";
 import { renderWithProviders } from "test-utils";
 import { describe, afterEach, vi, it, expect } from "vitest";
-import { ExplorerTree } from "#/components/file-explorer/explorer-tree";
+import { ExplorerTree } from "#/components/explorer-tree";
 
 const FILES = ["file-1-1.ts", "folder-1-2"];
